% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceAtAge.R
\name{abundanceAtAge}
\alias{abundanceAtAge}
\title{Estimate of abundance at age}
\usage{
abundanceAtAge(data, year, age)
}
\arguments{
\item{data}{data.table formatted as explained above}

\item{year}{survey year to calculate abundance for}

\item{age}{to calculate abundance for}
}
\value{
estimated abundance for age.
}
\description{
Estimate of abundance at age
}
