% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceAtReader.R
\name{abundanceAtReader}
\alias{abundanceAtReader}
\title{Estimate of abundance for all ages for each reader
Essentially extracting data corresponding to each reader, but agesamplecount needs to
be corrected with respect to how many that are read for the specific reader.}
\usage{
abundanceAtReader(data, year, Plot = TRUE)
}
\arguments{
\item{data}{data.table formatted as explained above}

\item{year}{survey year to calculate abundance for}
}
\value{
estimated abundance for age for each reader.
}
\description{
Estimate of abundance for all ages for each reader
Essentially extracting data corresponding to each reader, but agesamplecount needs to
be corrected with respect to how many that are read for the specific reader.
}
