% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stratum_estimate.R
\name{calc_stratum_estimate}
\alias{calc_stratum_estimate}
\title{Calculate within each stratum estimate of number of fish within each age group}
\usage{
calc_stratum_estimate(fitobj, weights, INDSUM, A)
}
\arguments{
\item{fitobj}{object fitted by stan}

\item{weights}{Weight within each stratum (D*A/n)}

\item{INDSUM}{indsum}
}
\value{
Matrix containing probabilities for error readings
}
\description{
Calculate within each stratum estimate of number of fish within each age group
}
