% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcM.R, R/calcM2.R
\name{calcM}
\alias{calcM}
\title{Calculation of transition probabilities for age error readings.}
\usage{
calcM(theta, Amin, Amax)

calcM(theta, Amin, Amax)
}
\arguments{
\item{theta}{(alpha0,alpha1,beta0,beta1,log(phi))}

\item{Amin}{mininum age}

\item{Amax}{maximum age}
}
\value{
Transition matrix

Transition matrix
}
\description{
Default values on parameters are from fitted model to ...

Default values on parameters are from fitted model to ...
}
