% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcM3.R
\name{calcM3}
\alias{calcM3}
\title{Calculation of transition probabilities for age error readings.}
\usage{
calcM3(theta, Amin, Amax)
}
\arguments{
\item{theta}{(alpha0,alpha1,beta0,beta1,log(phi))}

\item{Amin}{mininum age}

\item{Amax}{maximum age}
}
\value{
Transition matrix
}
\description{
Default values on parameters are from fitted model to ...
}
