% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateM.R
\name{calculateM}
\alias{calculateM}
\title{Calculation of transition probabilities for age error readings.}
\usage{
calculateM(
  beta0 = 4.699026,
  beta1 = -0.3507445,
  alpha0 = 0.7443011,
  alpha1 = -0.0381708,
  phi = 0.325192
)
}
\arguments{
\item{beta0}{beta0}

\item{beta1}{beta1}

\item{alpha0}{alpha0}

\item{alpha1}{alpha1}

\item{phi}{phi}
}
\value{
Transition matrix
}
\description{
Default values on parameters are from fitted model to ...
}
