% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNA.R
\name{calNA}
\alias{calNA}
\title{Calculation of transition probabilities for age error readings.}
\usage{
calNA(w, xiks, stratum, Y, K, S, A)
}
\arguments{
\item{w}{An y times K matrix giving}

\item{Amin}{mininum age}

\item{Amax}{maximum age}
}
\value{
Transition matrix
}
\description{
Default values on parameters are from fitted model to ...
}
