% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_m.R
\name{fitted_m}
\alias{fitted_m}
\title{Calculates matrix for reading errors given a set of parameters}
\usage{
fitted_m(para)
}
\arguments{
\item{para}{A vector of length 5 with (beta_0,beta_1,alpha_0,alpha_1,log(phi))}

\item{Amin}{Mininum age}

\item{Amax}{Maximum age}
}
\value{
Matrix containing probabilities for error readings
}
\description{
Calculates matrix for reading errors given a set of parameters
}
