% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Numfish_sample.R
\name{Make_Numfish_sample}
\alias{Make_Numfish_sample}
\title{Calculate samples of number of fish based on posterior samples of xi1
and boostrap samples of abundance}
\usage{
Make_Numfish_sample(xi1, bootD, stratayearindex)
}
\arguments{
\item{xi1}{Array of dimension MxKxA giving posterior samples of xi1, the age distribution per strata}

\item{bootD}{bootstrap samples of abundance within each strata}
}
\value{
estimated abundance for age for each reader.
}
\description{
Calculate samples of number of fish based on posterior samples of xi1
and boostrap samples of abundance
}
