% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ring.readErr.R
\docType{data}
\name{Ring.readErr}
\alias{Ring.readErr}
\title{Ring data error readings}
\format{
## `Ring.readErr`
An array of dimension  6x13x13 giving for each reader r=1,...,6 the number of true compared to read ages
}
\source{
"HI"
}
\usage{
Ring.readErr
}
\description{
Generamted by Make_ring_data.R
}
\keyword{datasets}
